requirements.txt
dnspython
folium
requests
tkinterweb
Pillow
